refactor!: Move split_*_epochs to allen vc ophys pipeline
